Exercise 3
Follow part 5 of the AddressApp tutorial which will teach you to persist the data between sessions, so that the contacts of your application is saved to the disk.
Notice: The tutorial claims that you can persist to XML out of the box, as the JAXB-library is embedded within the JDK. While this was true when the tutorial was written, but today this library has been separated into it's own library. Therefore you need to add three more Maven-dependencies to your project:
javax.xml.bind:jaxb-api:2.3.0
javax.activation:activation:1.1
org.glassfish.jaxb:jaxb-runtime:2.3.0